model/multimodal/requirements.txt
beautifulsoup4==4.13.5
boto3==1.40.31
faiss_cpu==1.12.0
fastapi==0.116.2
faster_whisper==1.2.0
fonttools==4.59.2
google_search_results==2.4.2
mammoth==1.10.0
numpy==2.3.3
openai==1.108.0
opencv_contrib_python==4.11.0.86
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
pdf2image==1.17.0
pdfplumber==0.11.7
Pillow==11.3.0
pvporcupine==3.0.5
PyAudio==0.2.14
pydub==0.25.1
pyhwp==0.1b15
pypdf==6.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
python_docx==1.2.0
Requests==2.32.5
sounddevice==0.5.2
soundfile==0.13.1
torch==2.8.0
torchaudio==2.8.0
transformers==4.56.0
webrtcvad==2.0.10
